+++ src/components/DropDown.tsx
@@ -4,7 +4,6 @@ import { DropdownType, HighlightType } from '../types/dropdown';
 import { ReactComponent as Union } from '../assets/union_icon.svg';
 import { ReactComponent as Spinner } from '../assets/spinner_icon.svg';
 import useIntersectionObserver from '../hooks/useIntersectionObserver';
-import { useSearchDispatch, useSearchState } from '../context/SearchProvider';
 
 function HighlightedText({ text, highlight }: HighlightType) {
   const parts = text.split(highlight);
@@ -21,10 +20,15 @@ function HighlightedText({ text, highlight }: HighlightType) {
   );
 }
 
-function DropDown({ dropdownRef, searchListData, inputText, handleAddTodoClick }: DropdownType) {
-  const { isTotal, isSearchLoading } = useSearchState();
-  const { handleSearchFetch } = useSearchDispatch();
-
+function DropDown({
+  dropdownRef,
+  searchListData,
+  inputText,
+  isTotal,
+  isSearchLoading,
+  handleAddTodoClick,
+  handleSearchFetch,
+}: DropdownType) {
   const onIntersect: IntersectionObserverCallback = ([{ isIntersecting }]) => {
     if (isIntersecting && !isTotal && !isSearchLoading) {
       handleSearchFetch('scroll', inputText);
